# docs/swagger.yaml
basePath: /api/
definitions:
...
  db.Reporte:
    properties:
      datos:
        items:
          type: integer
        type: array
      fechaGeneracion:
        $ref: '#/definitions/pgtype.Timestamp'
      generadoPor:
        type: integer
      periodoFin:
        $ref: '#/definitions/pgtype.Date'
      periodoInicio:
        $ref: '#/definitions/pgtype.Date'
      reporteID:
        type: integer
      tipoReporte:
        type: string
    type: object
...
  handler.CreateReporteRequest:
    type: object
  handler.CreateReporteResponse:
    properties:
      reporte_id:
        type: integer
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: Descriptive error message
        type: string
    type: object
...
  handler.UpdateReporteRequest:
    type: object
...
  pgtype.Date:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.InfinityModifier:
    enum:
    - 1
    - 0
    - -1
    type: integer
    x-enum-varnames:
    - Infinity
    - Finite
    - NegativeInfinity
...
  pgtype.Timestamp:
    properties:
      infinityModifier:
        $ref: '#/definitions/pgtype.InfinityModifier'
      time:
        description: Time zone will be ignored when encoding to PostgreSQL.
        type: string
      valid:
        type: boolean
    type: object
host: matwa.tail013c29.ts.net
info:
  contact: {}
  description: This is the API for my project.
  termsOfService: http://swagger.io/terms/
  title: Proyecto de Ingenieria de  Datos API
  version: "1.0"
paths:
...
  /v1/reportes:
    get:
      description: Retrieves reports filtered by date period.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: periodo_inicio
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: periodo_fin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved reportes
          schema:
            items:
              $ref: '#/definitions/db.Reporte'
            type: array
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to retrieve reportes
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: List Reportes by Period
      tags:
      - Reportes
    post:
      consumes:
      - application/json
      description: Creates a new report.
      parameters:
      - description: Reporte Data
        in: body
        name: reporte
        required: true
        schema:
          $ref: '#/definitions/handler.CreateReporteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created reporte
          schema:
            $ref: '#/definitions/handler.CreateReporteResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to create reporte
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create Reporte
      tags:
      - Reportes
  /v1/reportes/{id}:
    delete:
      description: Deletes a report by its ID.
      parameters:
      - description: Reporte ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Successfully deleted reporte
        "400":
          description: Invalid reporte ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to delete reporte
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Reporte
      tags:
      - Reportes
    get:
      description: Retrieves a specific report by its ID.
      parameters:
      - description: Reporte ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved reporte
          schema:
            $ref: '#/definitions/db.Reporte'
        "400":
          description: Invalid reporte ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Reporte not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to retrieve reporte
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get Reporte by ID
      tags:
      - Reportes
    put:
      consumes:
      - application/json
      description: Updates the data of an existing report.
      parameters:
      - description: Reporte ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Reporte Data
        in: body
        name: reporte
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateReporteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated reporte
          schema:
            $ref: '#/definitions/db.Reporte'
        "400":
          description: Invalid request body or reporte ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Reporte not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Failed to update reporte
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update Reporte
      tags:
      - Reportes
...
schemes:
- https
swagger: "2.0"